+++ langchain/src/agents/agent_toolkits/zapier/zapier.ts
@@ -1,7 +1,6 @@
 import { Toolkit } from "../base.js";
 import { Tool } from "../../tools/index.js";
-import { ZapierNLAWrapper } from "../../../zapier.js";
-import { ZapierNLARunAction } from "../../tools/zapier.js";
+import { ZapierNLARunAction, ZapierNLAWrapper } from "../../tools/zapier.js";
 
 export class ZapierToolKit extends Toolkit {
   tools: Tool[] = [];
