+++ langchain/src/agents/tools/index.ts
@@ -21,3 +21,4 @@ export {
 } from "./json.js";
 export { RequestsGetTool, RequestsPostTool } from "./requests.js";
 export { VectorStoreQATool } from "./vectorstore.js";
+export { ZapierNLARunAction, ZapierNLAWrapper } from "./zapier.js";
